refactor: move definitions for Modifier
